Refactor overlay configuration UI into module
